fix: splashExample 앱 depedency 오류 해결
